# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Qatar University Advisor API
  description: |
    واجهة برمجة التطبيقات للمستشار الجامعي الذكي — دولة قطر

    API for the Qatar University Advisor chatbot, serving students in Qatar
    via WhatsApp and web interface.

    شركة أذكياء للبرمجيات | Al-Nukhbawiya Software Company
  version: 1.3.0
  contact:
    name: شركة أذكياء للبرمجيات
    url: https://qatar-university-advisor.vercel.app
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary

servers:
  - url: https://qatar-university-advisor.vercel.app/api
    description: Production (Vercel)
  - url: http://localhost:3000/api
    description: Local development

paths:
  /webhook:
    get:
      operationId: verifyWebhook
      summary: التحقق من Webhook — WhatsApp Verification
      description: |
        يُستخدم من Meta/WhatsApp للتحقق من ملكية الـ webhook عند الإعداد الأولي.
        يتطلب مطابقة hub.verify_token مع WEBHOOK_VERIFY_TOKEN.
      tags:
        - WhatsApp
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
            enum: [subscribe]
          description: يجب أن يكون "subscribe"
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
          description: رمز التحقق المطابق لـ WEBHOOK_VERIFY_TOKEN
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
          description: رمز التحدي المُعاد لتأكيد الملكية
      responses:
        '200':
          description: تم التحقق بنجاح — يُعاد hub.challenge
          content:
            text/plain:
              schema:
                type: string
              examples:
                verified:
                  value: "1234567890"
                  summary: hub.challenge المُعاد
        '403':
          description: فشل التحقق — الرمز غير مطابق
          content:
            text/plain:
              schema:
                type: string
              examples:
                forbidden:
                  value: "Forbidden"

    post:
      operationId: handleIncomingMessage
      summary: استقبال رسائل WhatsApp الواردة
      description: |
        يستقبل رسائل WhatsApp من Meta Cloud API ويعالجها.

        الأمان:
        - التحقق من توقيع X-Hub-Signature-256 (HMAC-SHA256)
        - Rate limiting: 60 طلب/دقيقة لكل IP
        - Dedup: تجاهل الرسائل المكررة
        - حد أقصى 4000 حرف للرسالة

        أنواع الرسائل المدعومة: text، interactive (button_reply، list_reply).
      tags:
        - WhatsApp
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
          description: توقيع HMAC-SHA256 من Meta — sha256=<hex>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entry:
                  type: array
                  items:
                    type: object
                    properties:
                      changes:
                        type: array
                        items:
                          type: object
                          properties:
                            value:
                              type: object
                              properties:
                                messages:
                                  type: array
                                  items:
                                    type: object
                                    properties:
                                      id:
                                        type: string
                                        description: معرف الرسالة الفريد
                                      from:
                                        type: string
                                        description: رقم هاتف المرسل
                                      type:
                                        type: string
                                        enum: [text, interactive, image, audio]
                                      text:
                                        type: object
                                        properties:
                                          body:
                                            type: string
                                      interactive:
                                        type: object
                                        properties:
                                          type:
                                            type: string
                                            enum: [button_reply, list_reply]
                                          button_reply:
                                            type: object
                                            properties:
                                              title:
                                                type: string
                                          list_reply:
                                            type: object
                                            properties:
                                              title:
                                                type: string
            examples:
              textMessage:
                summary: رسالة نصية عادية
                value:
                  entry:
                    - changes:
                        - value:
                            messages:
                              - id: "wamid.abc123"
                                from: "97412345678"
                                type: "text"
                                text:
                                  body: "جامعة قطر"
              buttonReply:
                summary: رد على زر تفاعلي
                value:
                  entry:
                    - changes:
                        - value:
                            messages:
                              - id: "wamid.def456"
                                from: "97412345678"
                                type: "interactive"
                                interactive:
                                  type: "button_reply"
                                  button_reply:
                                    title: "جميع الجامعات"
      responses:
        '200':
          description: تم الاستلام والمعالجة بنجاح
          content:
            text/plain:
              schema:
                type: string
              examples:
                ok:
                  value: "OK"
        '400':
          description: خطأ في تحليل JSON
          content:
            text/plain:
              schema:
                type: string
              examples:
                badRequest:
                  value: "Bad Request"
        '403':
          description: فشل التحقق من التوقيع
          content:
            text/plain:
              schema:
                type: string
              examples:
                forbidden:
                  value: "Forbidden"
        '415':
          description: Content-Type غير مدعوم (يجب أن يكون application/json)
          content:
            text/plain:
              schema:
                type: string
              examples:
                unsupported:
                  value: "Unsupported Media Type"
        '429':
          description: تجاوز حد الطلبات — Rate limit exceeded
          content:
            text/plain:
              schema:
                type: string
              examples:
                rateLimit:
                  value: "Too Many Requests"
        '405':
          description: HTTP method غير مدعوم
          content:
            text/plain:
              schema:
                type: string
              examples:
                methodNotAllowed:
                  value: "Method Not Allowed"

  /health:
    get:
      operationId: healthCheck
      summary: فحص حالة النظام — Health Check
      description: |
        يفحص حالة جميع الخدمات: Supabase، الذاكرة، Redis.

        - 200: النظام سليم (healthy)
        - 503: النظام متدهور أو معطل (degraded/unhealthy)

        Rate limit: 10 طلبات/دقيقة.
      tags:
        - Monitoring
      responses:
        '200':
          description: النظام يعمل بشكل سليم
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                healthy:
                  summary: جميع الخدمات تعمل
                  value:
                    status: "healthy"
                    timestamp: "2026-04-04T12:00:00.000Z"
                    version: "1.2.0"
                    uptime: 86400
                    services:
                      supabase:
                        status: "healthy"
                        latency: 45
                      memory:
                        status: "healthy"
                        used: "120MB"
                        total: "256MB"
                        rss: "180MB"
                      redis:
                        status: "skipped"
        '405':
          description: Method not allowed — فقط GET مسموح
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                notAllowed:
                  value:
                    error: "Method not allowed"
        '429':
          description: Rate limit exceeded
        '503':
          description: النظام متدهور — خدمة واحدة أو أكثر معطلة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                degraded:
                  summary: Supabase معطلة
                  value:
                    status: "degraded"
                    timestamp: "2026-04-04T12:00:00.000Z"
                    version: "1.2.0"
                    uptime: 86400
                    services:
                      supabase:
                        status: "unhealthy"
                        error: "Connection failed"
                        latency: 5000
                      memory:
                        status: "healthy"
                        used: "120MB"
                        total: "256MB"
                        rss: "180MB"
                      redis:
                        status: "skipped"

  /admin:
    get:
      operationId: getAdminDashboard
      summary: لوحة تحكم المدير — Admin Dashboard
      description: |
        تعرض إحصائيات النظام وأكثر الاستفسارات شيوعاً وحالة الخدمات.

        الأمان:
        - Basic Auth (Bearer token) — يتطلب ADMIN_PASSWORD
        - CORS مقيد بـ ADMIN_ORIGIN
        - Rate limit: 5 طلبات/دقيقة
        - SEC-001: مقارنة كلمة المرور بـ timing-safe comparison
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: بيانات لوحة التحكم
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
              examples:
                dashboard:
                  summary: بيانات لوحة التحكم
                  value:
                    timestamp: "2026-04-04T12:00:00.000Z"
                    stats:
                      totalUsers: 1250
                      totalMessages: 45000
                      totalQueries: 38000
                    topQueries:
                      - query: "جامعة قطر"
                        count: 520
                      - query: "الابتعاث الأميري"
                        count: 310
                      - query: "الكليات العسكرية"
                        count: 280
                    circuitStatus:
                      isHealthy: true
                      state: "CLOSED"
                      failures: 0
                    botStatus:
                      vercel: "operational"
                      whatsapp: "operational"
                      gemini: "operational"
                      supabase: "operational"
        '401':
          description: غير مصرح — كلمة المرور خاطئة أو مفقودة
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                unauthorized:
                  value:
                    error: "Unauthorized"
        '405':
          description: Method not allowed — فقط GET مسموح
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                notAllowed:
                  value:
                    error: "Method not allowed"
        '429':
          description: Rate limit exceeded
        '500':
          description: خطأ داخلي
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '503':
          description: ADMIN_PASSWORD غير مُعدّ
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                notConfigured:
                  value:
                    error: "Admin panel not configured. Set ADMIN_PASSWORD environment variable."

  /cron-pdppl-cleanup:
    get:
      operationId: pdpplCleanupCron
      summary: تنظيف بيانات PDPPL — Cron Job
      description: |
        وظيفة مجدولة تعمل يومياً لحذف/إخفاء هوية المستخدمين وفق قانون PDPPL القطري.

        العمليات:
        1. حذف المستخدمين الذين طلبوا الحذف (بعد 7 أيام grace period)
        2. إخفاء هوية المستخدمين المنتهية فترة احتفاظ بياناتهم
        3. تسجيل النتائج في جدول analytics

        الأمان: يتطلب CRON_SECRET في header Authorization.
      tags:
        - PDPPL Compliance
      security:
        - cronAuth: []
      responses:
        '200':
          description: اكتمل التنظيف بنجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CronCleanupResponse'
              examples:
                success:
                  summary: تم حذف وإخفاء هوية مستخدمين
                  value:
                    success: true
                    deleted: 3
                    anonymized: 7
                    errors: []
                    timestamp: "2026-04-04T02:00:00.000Z"
                    message: "PDPPL cleanup complete: 3 deleted, 7 anonymized"
                noAction:
                  summary: لا يوجد مستخدمون للتنظيف
                  value:
                    success: true
                    deleted: 0
                    anonymized: 0
                    errors: []
                    timestamp: "2026-04-04T02:00:00.000Z"
                    message: "PDPPL cleanup complete: 0 deleted, 0 anonymized"
        '401':
          description: غير مصرح — CRON_SECRET غير مطابق
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                unauthorized:
                  value:
                    error: "Unauthorized"
        '405':
          description: Method not allowed — فقط GET و POST مسموحان
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                notAllowed:
                  value:
                    error: "Method not allowed"
        '500':
          description: خطأ فادح أثناء التنظيف
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  deleted:
                    type: integer
                  anonymized:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
                  timestamp:
                    type: string
                    format: date-time
        '503':
          description: قاعدة البيانات غير مُعدَّة
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                notConfigured:
                  value:
                    error: "Database not configured"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        كلمة مرور المدير — ADMIN_PASSWORD
        Authorization: Bearer <password>
    cronAuth:
      type: http
      scheme: bearer
      description: |
        رمز التشغيل المجدول — CRON_SECRET
        Authorization: Bearer <cron_secret>

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - version
        - uptime
        - services
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          description: الحالة العامة للنظام
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          description: إصدار التطبيق من package.json
        uptime:
          type: integer
          description: وقت التشغيل بالثواني
        services:
          type: object
          properties:
            supabase:
              type: object
              properties:
                status:
                  type: string
                  enum: [healthy, unhealthy, not_configured]
                latency:
                  type: integer
                  description: زمن الاستجابة بالمللي ثانية
                error:
                  type: string
            memory:
              type: object
              properties:
                status:
                  type: string
                  enum: [healthy, warning]
                used:
                  type: string
                total:
                  type: string
                rss:
                  type: string
            redis:
              type: object
              properties:
                status:
                  type: string
                  enum: [healthy, unhealthy, skipped]

    AdminResponse:
      type: object
      required:
        - timestamp
        - stats
        - topQueries
        - circuitStatus
        - botStatus
      properties:
        timestamp:
          type: string
          format: date-time
        stats:
          type: object
          properties:
            totalUsers:
              type: integer
            totalMessages:
              type: integer
            totalQueries:
              type: integer
        topQueries:
          type: array
          items:
            type: object
            properties:
              query:
                type: string
              count:
                type: integer
        circuitStatus:
          type: object
          properties:
            isHealthy:
              type: boolean
            state:
              type: string
              enum: [CLOSED, OPEN, HALF_OPEN]
            failures:
              type: integer
        botStatus:
          type: object
          properties:
            vercel:
              type: string
              enum: [operational, degraded, down]
            whatsapp:
              type: string
              enum: [operational, degraded, down]
            gemini:
              type: string
              enum: [operational, degraded, down]
            supabase:
              type: string
              enum: [operational, degraded, down]

    CronCleanupResponse:
      type: object
      required:
        - success
        - deleted
        - anonymized
        - errors
        - timestamp
        - message
      properties:
        success:
          type: boolean
        deleted:
          type: integer
          description: عدد المستخدمين المحذوفين
        anonymized:
          type: integer
          description: عدد المستخدمين مخفيي الهوية
        errors:
          type: array
          items:
            type: string
          description: أي أخطاء حدثت أثناء التنظيف
        timestamp:
          type: string
          format: date-time
        message:
          type: string
          description: ملخص عملية التنظيف

tags:
  - name: WhatsApp
    description: معالجة رسائل WhatsApp الواردة والصادرة
  - name: Monitoring
    description: مراقبة حالة النظام والخدمات
  - name: Admin
    description: لوحة تحكم المدير — إحصائيات وإدارة
  - name: PDPPL Compliance
    description: |
      الامتثال لقانون حماية البيانات الشخصية القطري رقم 13 لسنة 2016